+++ app/resources/en/translation/documents.json
@@ -3,24 +3,23 @@
   "header": "Send your documents",
   "heading1": "1. Prepare to send your documents",
   "heading2": "2. Send your documents to the probate registry",
-  "text1": "If you don&rsquo;t have a printer, write your reference number {ccdReferenceNumber} on a sheet of paper and include it with your application.",
-  "text2": "You need to send the probate registry:",
-  "text3": "You&rsquo;ll need to include a cover letter if the will or codicils have changed in any way since you&rsquo;ve had them.",
-  "text4": "The letter should explain whats been changed and why.",
-  "text5": "You should send your documents using a signed-for postal service to:",
-  "warning": "The grant of probate won&rsquo;t be issued if you don&rsquo;t have all the necessary documents.",
-  "address": "Digital Application\nOxford District Probate Registry\nCombined Court Building\nSt Aldates\nOxford\nOX1 1LY",
-  "checklist-item1-no-codicils": "the original will (this won&rsquo;t be returned to you)",
-  "checklist-item1-codicils": "the original will and any updates (codicils) (these won&rsquo;t be returned to you)",
-  "checklist-item2-spouse-renouncing": "a <a href=\"{renunciationFormLink}\" class=\"govuk-link\" target=\"_blank\" rel=\"external noopener\" aria-label=\"renunciation form (opens in a new window)\">renunciation form (opens in a new window)</a> filled in by the spouse or civil partner of the deceased who is permanently giving up the right to make this application for probate",
-  "checklist-item3-will-uploaded": "the original death certificate (this will be returned to you)",
-  "checklist-item4-iht205": "the completed inheritance tax form IHT 205",
-  "checklist-item5-renunciated": "a <a href=\"{renunciationFormLink}\" target=\"_blank\" class=\"govuk-link\" rel=\"external noopener\" aria-label=\"renunciation form (opens in a new window)\">renunciation form (opens in a new window)</a> filled in by any executors who are permanently giving up their right to make this application for probate",
-  "checklist-item6-deed-poll": "copy of the deed poll document for {executorCurrentName}",
-  "checkboxLabel": "I understand that I need to sign the will before sending my documents",
-  "checkboxLabel-codicils": "I understand that I need to sign the will and any &lsquo;codicils&rsquo; before sending my documents",
-  "coverSheetDownloadMessage": "Download and print this",
-  "coverSheetPdf": "application coversheet",
+  "text1": "You must send the probate registry:",
+  "text2": "Send your documents using a signed postal service to:",
+  "warning": "We cannot give you a grant of probate without the required documents.",
+  "address": "Principal Registry of the Family Division (PRFD)\nHMCTS Probate\nPO BOX 12625\nHarlow\nCM20 9QE",
+  "checklist-item1-application-coversheet": "either your completed <a class=\"govuk-link\" id=\"coverSheetPdfHref\" href=\"/cover-sheet-pdf\">application coversheet</a> or write your reference number {ccdReferenceNumber} on a sheet of paper",
+  "checklist-item2-no-codicils": "the original will (by law, we must keep your original will as it becomes a public document)",
+  "checklist-item2-codicils": "the original will and any codicils (by law, we must keep your original will as it becomes a public document)",
+  "checklist-item3-will-damage-no-codicils": "if the will appears to have had something added, removed, or is damaged you will need to write a letter to explain. Damage could include staple or punch holes, rusty paper clip marks, torn edges, water damage etc.",
+  "checklist-item3-will-damage-codicils": "if the will or codicil appears to have had something added, removed, or is damaged you will need to write a letter to explain. Damage could include staple or punch holes, rusty paper clip marks, torn edges, water damage etc.",
+  "checklist-item4-interim-death-cert": "the interim death certificate",
+  "checklist-item4-foreign-death-cert": "the original foreign death certificate",
+  "checklist-item4-foreign-death-cert-translation": "the English translation of the foreign death certificate",
+  "checklist-item5-foreign-death-cert-PA19": "either a completed PA19 - <a target=\"_blank\" rel=\"external noopener\" href=\"{applicationFormPA19}\" class=\"govuk-link\">Foreign death certificate form (opens in a new window)</a> or a certificate from the translator confirming the translation's validity",
+  "checklist-item6-spouse-renouncing": "a <a href=\"{renunciationFormLink}\" class=\"govuk-link\" target=\"_blank\" rel=\"external noopener\" aria-label=\"renunciation form (opens in a new window)\">renunciation form (opens in a new window)</a> filled in by the spouse or civil partner of the deceased who is permanently giving up the right to make this application for probate",
+  "checklist-item7-iht205": "the completed inheritance tax form IHT 205",
+  "checklist-item8-renunciated": "a <a href=\"{renunciationFormLink}\" target=\"_blank\" class=\"govuk-link\" rel=\"external noopener\" aria-label=\"renunciation form (opens in a new window)\">renunciation form (opens in a new window)</a> filled in by any executors who are permanently giving up their right to make this application for probate",
+  "checklist-item9-deed-poll": "copy of the deed poll document for {executorCurrentName}",
 
   "errors": {
     "sentDocuments": {
